Move `guess_file_type` into `hermes.utils`
